+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 3.4.0 - 2026-06-18
 
 * Added: Support for faster dot-sourcing module loading and module debugging mode
 * Added: PowerShell module script (psm1) tests for module behavior, metadata and loader